Fixed infill combination on Prusa XL with different sized nozzles
Previously infill combination on Prusa XL only combines layers up to the height of the smaller nozzle out of solid infill extruder and infill extruder.

Solid infills are usually on the surface and are usually of a smaller nozzle for better quality. However, internal sparse infills are always internal and are usually printed with a larger nozzle.

This commit now combines internal sparse infills up to the size of its nozzle size, while leaving solid infill untouched. Which significantly speeds up prints on Prusa XL with different sized nozzles installed.